Pulled out the helpers (translation and session)
