+++ helpers/session_helper.rb
@@ -1,0 +1,5 @@
+module SessionHelper
+  def current_user
+    User[session.fetch('User')] || NullUser.new
+  end
+end
