+++ helpers/translations_helper.rb
@@ -1,0 +1,9 @@
+module TranslationsHelper
+  def t(key)
+    translator.translate(key)
+  end
+
+  def translator
+    @translator ||= Translator.new(:de)
+  end
+end
